requirements.txt
asgiref==3.7.2
coverage==7.3.4
distlib==0.3.7
dj-database-url==2.1.0
Django==4.2.5
django-heroku==0.3.1
filelock==3.13.1
gunicorn==21.2.0
mysqlclient==2.2.0
packaging==23.2
Pillow==10.0.1
platformdirs==4.1.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
virtualenv==20.25.0
whitenoise==6.6.0
